Fixed wrong PATH for ready module, when using standalone build.
Thanks to Chris Clayton <chris2553@googlemail.com>

he sent me this.

--- exfat/Makefile      2013-11-30 09:49:05.451945664 +0000
+++ exfat/Makefile.fixed        2014-01-05 09:13:22.643866902 +0000
@@ -17,6 +17,7 @@ EXTRA_FLAGS += -I$(PWD)

 #KDIR  := /usr/src/linux/
 KDIR   := /lib/modules/$(shell uname -r)/build
+MDIR   := /lib/modules/$(shell uname -r)
 PWD            := $(shell pwd)

 export CONFIG_EXFAT_FS := m
@@ -31,12 +32,12 @@ help:
        $(MAKE) -C $(KDIR) M=$(PWD) help

 install: exfat.ko
-       rm -f ${KDIR}/../kernel/fs/exfat/exfat.ko
-       install -m644 -b -D exfat.ko ${KDIR}/../kernel/fs/exfat/exfat.ko
+       rm -f ${MDIR}/kernel/fs/exfat/exfat.ko
+       install -m644 -b -D exfat.ko ${MDIR}/kernel/fs/exfat/exfat.ko
        depmod -aq

 uninstall:
-       rm -rf ${KDIR}/../kernel/fs/exfat
+       rm -rf ${MDIR}/kernel/fs/exfat
        depmod -aq

 endif

I find that the module is installed in the wrong place.
For example, if the exfat source is located in $HOME/kernel/exfat,
the module is installed as $HOME/kernel/kernel/fs/exfat/exfat.ko.
This is because KDIR refers to the module's source directory,
not the directory in which the kernel and its modules are installed.

The patch below fixes this.